[Android] Change the API for ApplicationStatus in base dir.
Chromium class inherits from Application directly. This is not suitable
for Crosswalk embedding API. Instead, define a new API which can pass Application
instance as the parameter instead of asking users' application inheriting from
Chromium Application.
This fix is due to the change in upstream:
https://codereview.chromium.org/159173002
